+++ modules/ipc/src/test/java/com/nautilus_technologies/tsubakuro/channel/ipc/sql/ResultSetTotalTest.java
@@ -14,11 +14,11 @@ import org.msgpack.value.ValueType;
 import com.nautilus_technologies.tsubakuro.channel.ipc.SessionWireImpl;
 import com.nautilus_technologies.tsubakuro.protos.ProtosForTest;
 import com.nautilus_technologies.tsubakuro.util.ByteBufferInputStream;
-import com.nautilus_technologies.tsubakuro.channel.common.ChannelResponse;
+import com.nautilus_technologies.tsubakuro.channel.common.connection.wire.ChannelResponse;
 import com.tsurugidb.jogasaki.proto.SqlResponse;
 
 class ResultSetTotalTest {
-    static final long SERVICE_ID_SQL = 3;
+    static final int SERVICE_ID_SQL = 3;
     private SessionWireImpl client;
     private ServerWireImpl server;
     private final String dbName = "tsubakuro";
